allowed_hosts_rsa_keys/authorized_keys: add allowed ssh keys from specific hosts
by running the following command from selected hosts:
yogev,yan,dror,build-servers

we have a list of hosts that are allowed to ssh into the tink OSIE
image.

in the future we should probably add more users to allow this ssh.

we run the following command:

ssh-copy-id light@rack03-server60

which in turn will add the id_rsa.pub of each client to the
authorized_keys and then we add this file to the image
by specifying it in hook.yaml