hwinfo: add ability to detect all net ifaces and assign data ip
we want to be able to assign data-ip for the data interfaces like
mlx and intel when we boot to osie.

this will allow us to communicate with pulp over faster network
then the access which is 1g only.

hwinfo will detect all ifaces, classify the ifaces as data/access
calculate the new ipv4 address using the following formula:

if access's IPv4 is A.B.C.D we will assign a new IP: 10.100.C.D

10.100 prefix is garentied to be unique for data network usage
C.D is unique by the DHCP and dnsmask.